add _form partial

+++ src/Controller/PinsController.php
@@ -94,14 +94,17 @@ class PinsController extends AbstractController
 
     /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="app_pins_delete", methods={"DELETE"})
+     * @param Request $request
      * @param Pin $pin
      * @param EntityManagerInterface $em
      * @return Response
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pins_delete' . $pin->getId(), $request->request->get('csrf_token'))){
+            $em->remove($pin);
+            $em->flush();
+        }
         return $this->redirectToRoute('app_home');
     }
